feat: Add support for systemd notify
